Simplify auth and refactoring
Simplifying auth by doing the following:
* removing the singleton, it was only needed when reading
from a file
* removing the hashing, we already fetch the secret in memory
cache and in the secret it's encoded, k8s takes care of giving
us the value so we can compare
* move auth to middleware as it was simplified
* clusterReady logic is simplified by splitting each component
into it's own method to make it clearer. Current definition
requires kubevirt/rke/longhorn readiness so they have their
own methods where "what ready means" can be seen simplified.
